refactor: extract title/artist vote-window cluster from state.py (#1271)
Pull the #1180 Phase 4 REVEAL vote-window scheduling + finalization
writers out of the GameState God-Object into a new VoteWindowMixin
(game/state_vote_window.py), following the established state_*.py mixin
pattern. This is the writer half the earlier ChallengeMixin cut
deliberately left behind (coupled to the score lock + auto-advance task).

Moved verbatim (behavior-preserving, API byte-identical):
- _schedule_title_artist_vote_window
- _title_artist_vote_window
- _finalize_title_artist_window
- _score_title_artist_round
- resolve_title_artist_if_pending

state.py: 2269 -> 2174 lines (-95). GamePhase is imported lazily inside
the window task (powerups/serializers pattern) to avoid a cyclic import;
TITLE_ARTIST_VOTE_WINDOW_SECONDS moves with the cluster.

The two window tests that monkeypatch the window-length constant now
patch it in its new home module (state_vote_window) — the patch target
moves with the symbol, no behavior change.

Refs #1271 (increment — issue stays OPEN).
